+++ python-lib/odata_client.py
@@ -2,6 +2,7 @@ import requests
 import logging
 from odata_constants import ODataConstants
 from dss_constants import DSSConstants
+from odata_common import get_odata_instance, get_list_title, get_login
 from dataikuapi.utils import DataikuException
 from time import sleep
 
@@ -16,15 +17,13 @@ class ODataClient():
 
     def __init__(self, config):
         self.auth_type = config.get(DSSConstants.AUTH_TYPE)
-        login = config.get('sap-odata_{}'.format(self.auth_type))
-        self.odata_service_node = config.get(ODataConstants.SERVICE_NODE).strip("/")
-        if self.odata_service_node != "":
-            self.odata_instance = "/".join([login[ODataConstants.INSTANCE].strip("/"), self.odata_service_node])
-        else:
-            self.odata_instance = login[ODataConstants.INSTANCE].strip("/")
-        self.ignore_ssl_check = login["ignore_ssl_check"]
-        self.odata_list_title = config.get(ODataConstants.LIST_TITLE)
-        odata_version = login[ODataConstants.VERSION]
+        login = get_login(config)
+        if not login:
+            raise Exception("Select a valid preset")
+        self.odata_instance = get_odata_instance(config)
+        self.ignore_ssl_check = login.get("ignore_ssl_check", False)
+        self.odata_list_title = get_list_title(config)
+        odata_version = login.get(ODataConstants.VERSION)
         self.set_odata_protocol_version(odata_version)
 
         if "sap-odata_oauth" in config and ODataConstants.OAUTH in config["sap-odata_oauth"]:
@@ -56,13 +55,20 @@ class ODataClient():
         session = requests.Session()
         if self.ignore_ssl_check is True:
             session.verify = False
-        login_config = config.get(ODataConstants.LOGIN)
+        login_config = config.get(ODataConstants.LOGIN, {}) if self.auth_type == "login" else config.get("sap-odata_user-account", {})
         if odata_version == ODataConstants.ODATA_VSAP:
+            if self.auth_type == "user-account":
+                username_password = login_config.get("username_password", {})
+                session.auth = (
+                    username_password.get("user", ""),
+                    username_password.get("password", "")
+                )
+            else:
+                session.auth = (
+                    login_config.get(ODataConstants.USERNAME, ""),
+                    login_config.get(ODataConstants.PASSWORD, "")
+                )
             self.sap_client = login_config.get(ODataConstants.SAP_CLIENT, "")
-            session.auth = (
-                login_config.get(ODataConstants.USERNAME, ""),
-                login_config.get(ODataConstants.PASSWORD, "")
-            )
             session.head(
                 self.odata_instance,
                 params={
@@ -78,10 +84,10 @@ class ODataClient():
             )
         return session
 
-    def get_entity_collections(self, entity, top=None, skip=None, page_url=None):
+    def get_entity_collections(self, entity="", top=None, skip=None, page_url=None, filter=None):
         if self.odata_list_title is None or self.odata_list_title == "":
             top = None  # SAP will complain if $top is present in a request to list entities
-        query_options = self.get_base_query_options(top=top, skip=skip)
+        query_options = self.get_base_query_options(top=top, skip=skip, filter=filter)
         url = page_url if page_url else self.odata_instance + '/' + entity.strip("/") + self.get_query_string(query_options)
         data = None
         while self._should_retry(data):
@@ -133,7 +139,7 @@ class ODataClient():
         headers["Authorization"] = self.get_authorization_bearer()
         return headers
 
-    def get_base_query_options(self, top=None, skip=None, records_limit=None):
+    def get_base_query_options(self, top=None, skip=None, records_limit=None, filter=None):
         if self.force_json and self.json_in_query_string:
             query_options = [DSSConstants.JSON_FORMAT]
         else:
@@ -146,6 +152,8 @@ class ODataClient():
             query_options.append(ODataConstants.SKIP.format(skip))
         if top:
             query_options.append(ODataConstants.TOP.format(top))
+        if filter:
+            query_options.append(ODataConstants.FILTER.format(filter))
         return query_options
 
     def format(self, item):