+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Version 1.0.3](https://github.com/dataiku/dss-plugin-sap-odata/releases/tag/v1.0.3) - Feature - 2023-10-20
+
+- Add filter field
+- Convert OData dates to DSS format
+- Add individual user credentials preset
+- Renamed SAP/ODATA login parameter to Service account
+- Dynamic selectors for service name and entity
+
 ## [Version 1.0.2](https://github.com/dataiku/dss-plugin-sap-odata/releases/tag/v1.0.2) - Feature and fix release - 2022-03-22
 
 - Add v4 compatibility